Create package for the collection of data Using the setuptools setup a package can be created for the collection of data from youtube. This could make sharing the code easier. Additionally, the code could be more modular which makes testing it easier.
All code from jupyter notebook into modules
Simple first step to move all code to own module containing
a single entity as namedtuple.

Namedtuple contains the fieldnames. This enforces that the conversion
of the response object of to the domain object contains all fields.
The dictionary writer retrieves the header if the file is empty
from the fields of the namedtuple.
Because it is a dictionaryWriter the named tuple is converted to a
dictionary before being used by the dictionary writer.
The formatting is kept as close as possible to pep-8

Try-except to handle key errors are changed to using the get-method
of the dictionary data structure.

Add a getting started notebook as a code example

Complete api configuration with api.conf file where you can store the
key else it will be asked using getpass