Refactor the functions to be more pythonic
Renamed variables and modified whitespace to stay in line with pep-8
convention.
Move youtube client to own module and make the client an input parameter.
Construction of the client should not be the concern of a function

Try-except to handle key errors are changed to using the get-method
of the dictionary data structure.

Removed result variables when they are not modified and return the
result of a api-call directly