Refactor package to contain single entity as namedtuple
Namedtuple contains the fieldnames. This enforces that the conversion
of the response object of to the domain object contains all fields.
The dictionary writer retrieves the header if the file is empty
from the fields of the namedtuple.
Because it is a dictionaryWriter the named tuple is converted to a
dictionary before being used by the dictionary writer.