Fix the package install to be used in notebook
The package dir seems to be required to make the test run only
our source code and not the libs found in the virtualenv if the
virtualenv is also in the DataCollection folder.
This also applies for the coverage

However, this seems to hinder the correct installation of the package.
Since this has higher priority, the change is reversed